templates: Use section template
Instead of section_title() shortcode, use section template to generate
H1 header and short page description.

The old way was calling section_title() from every single sections. If
every section needs it, we can write it in the template. This commit
does just that.

section_title() is still used by sc_obc. Thus, it's not removed, yet.

There is also section_title partial template, which is still used by
news_section.html. This commit doesn't address it.

With this change, cookies are removed.  Both pages and sections know
where they are. You can get the location info from many variables with
out having it in the page.

  - section.permalink
  - section.path
  - section.components
  - section.slug

If we need to re-implement cookies, use one of them and implement in
templates/section.html

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>