templates: merge twocol shortcode into section template
twocol shortcode sets the section's background image. The name is
misleading due to historical reasons. This commit moves the shortcode
to the section template because every single section is using it.

There are no remaining users of the twocol shortcode, so it is also
removed in this commit.

CSS is not touched in this commit and needs to be addressed later.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>