content recruit: software: Add position overview
Introduce an overview section outlining the responsibilities and scope
of the Software Engineer role. Highlight key areas such as low-level
software development (OS layers, device drivers), use of Zephyr RTOS
and Linux, and integration of middlewares like NASA cFS and
libcsp. Emphasize the mix of in-house product development and client
project involvement, along with the collaborative nature of the
position.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>